Extract shared controller behaviour for managing home page lists
Use a similar technique to that for the shared model behaviour, but we need to be able to provide overrides.  Have used define_method rather than simple def for the methods as without it the local variables used to customize the method calls in the controller methods are not available.